refactor(common) : 공통 DTO 패키지 분리
